requirements.txt
# ResumeSearch 项目依赖

# OpenAI API
openai>=1.0.0

# Flask Web框架
Flask>=2.3.0
Flask-CORS>=4.0.0

# YAML配置文件解析
PyYAML>=6.0

# 其他工具
python-dotenv>=1.0.0
requests>=2.31.0
